+++ router/http/src/main/scala/io/buoyant/router/Http.scala
@@ -7,6 +7,7 @@ import com.twitter.finagle.param.{ProtocolLibrary, ResponseClassifier}
 import com.twitter.finagle.server.StackServer
 import com.twitter.finagle.service.StatsFilter
 import com.twitter.finagle.{Http => FinagleHttp, Server => FinagleServer, http => fhttp, _}
+import io.buoyant.router.ClassifiedRetries.ResponseDiscarder
 import io.buoyant.router.Http.param.HttpIdentifier
 import io.buoyant.router.http._
 import java.net.SocketAddress
@@ -38,10 +39,18 @@ object Http extends Router[Request, Response] with FinagleServer[Request, Respon
       .transformed(_.replace(http.TracingFilter.role, http.TracingFilter.module))
       .transformed(_.remove(TlsFilter.role))
 
+    val responseDiscarder = ResponseDiscarder[Response] { rsp =>
+      if (rsp.isChunked) {
+        rsp.reader.discard()
+      }
+    }
+    implicit val discarderParam = ResponseDiscarder.param[Response]
+
     val defaultParams: Stack.Params =
       StackRouter.defaultParams +
         fhttp.param.Streaming(true) +
         fhttp.param.Decompression(false) +
+        responseDiscarder +
         ProtocolLibrary("http")
   }