+++ linkerd/protocol/http/src/integration/scala/io/buoyant/linkerd/protocol/TlsStaticValidationTest.scala
@@ -47,6 +47,7 @@ class TlsStaticValidationTest extends FunSuite with Awaits {
                 await(client(req))
               }
               assert(rsp.contentString == "woof")
+              ()
 
             } finally await(client.close())
           } finally await(server.close())