+++ namerd/iface/interpreter-thrift/src/test/scala/io/buoyant/namerd/iface/ThriftNamerEndToEndTest.scala
@@ -20,7 +20,7 @@ class ThriftNamerEndToEndTest extends FunSuite with Eventually with IntegrationP
     interval = scaled(Span(100, Milliseconds))
   )
 
-  def retryIn() = 1.second
+  val retryIn = () => 1.second
   val clientId = Path.empty
   val ns = "testns"